parse bool into int

+++ src/main.rs
@@ -173,10 +173,10 @@ async fn post_status(pool_id: cynic::Id) {
                         let mut outside_putts = Vec::from(vec!["Outside putts::".to_string(), "".to_string(), "".to_string(), "".to_string()]);
                         for result in player.results {
                             personal_vec.push(result.score.to_string());
-                            ob_s.push(result.is_out_of_bounds.to_string());
-                            circle_hits.push(result.is_inside_putt.to_string());
-                            inside_putts.push(result.is_inside_putt.to_string());
-                            outside_putts.push(result.is_outside_putt.to_string());
+                            ob_s.push((result.is_out_of_bounds as u8).to_string());
+                            circle_hits.push((result.is_inside_putt as u8).to_string());
+                            inside_putts.push((result.is_inside_putt as u8).to_string());
+                            outside_putts.push((result.is_outside_putt as u8).to_string());
                             
                             
                         }
